requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
#pillow==10.3.0
#prettyprint==0.1.5
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.1
django-environ